% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athlete.R
\name{athlete}
\alias{athlete}
\title{Data for specific athlete}
\usage{
athlete(id)
}
\arguments{
\item{id}{ID for athlete}
}
\value{
A list with the following fields: \code{name}, \code{gender} and \code{results}.
}
\description{
Data for specific athlete
}
\examples{
athlete("7abaea70-337a-4c7a-9d13-31306dca26a8")
athlete("7ef6fbc8-4169-4a98-934e-ff5fa79ba103")
}
