% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_athlete}
\alias{search_athlete}
\title{Search for an athlete by name}
\usage{
search_athlete(name)
}
\arguments{
\item{name}{A regular expression}
}
\value{
A data frame with information on all athletes whose names matches regular expression.
}
\description{
Search for an athlete by name
}
\examples{
search_athlete("Mann$")
}
